Fix to git deploy script.  Fixed default path in CommitDatabaseScripts.  Fixed extractResourceScripts to work with folder based scripts

+++ GitHooks/deploy-database-git-scripts.ps1
@@ -155,7 +155,9 @@ function EnsureInsideRepo {
 Remove $line (entire) from file at $path
 #>
 function RemoveLine($line, $path) {
-    (Get-Content -Path $path) | Where-Object { $_ -ne $line } | Set-Content -Path $path
+    if (Test-Path $path) {
+        (Get-Content -Path $path) | Where-Object { $_ -ne $line } | Set-Content -Path $path
+    }
 }
 
 <#
